refactor(app-opine): move routing into separate opine Router instance
this will enable pulling in just the router, which is able to merge route
params, instead of the entire opine instance
